pyproject.toml
[project]
name = "obsidian-to-anki"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "html5lib>=1.1",
    "lxml>=6.0.1",
    "mistune>=3.1.4",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "pytest>=8.4.2",
]
